Support filament v3
