+++ composer.json
@@ -18,7 +18,7 @@
     "require": {
         "php": "^8.0 || ^8.1 || ^8.2",
         "spatie/laravel-package-tools": "^1.9.2",
-        "filament/filament": "^2"
+        "filament/filament": "^3"
     },
     "require-dev": {
         "laravel/pint": "^1.0",
