+++ src/FilamentSparkPieServiceProvider.php
@@ -2,19 +2,57 @@
 
 namespace AAbosham\FilamentSparkPie;
 
-use Filament\PluginServiceProvider;
+use Filament\Support\Assets\Asset;
+use Filament\Support\Assets\Js;
+use Filament\Support\Facades\FilamentAsset;
 use Spatie\LaravelPackageTools\Package;
+use Spatie\LaravelPackageTools\PackageServiceProvider;
 
-class FilamentSparkPieServiceProvider extends PluginServiceProvider
+class FilamentSparkPieServiceProvider extends PackageServiceProvider
 {
-    protected array $beforeCoreScripts = [
-        'spark-pie' => __DIR__ . '/../resources/dist/js/spark-pie.js',
-    ];
+    public static string $name = 'filament-spark-pie';
+
+    public static string $viewNamespace = 'filament-spark-pie';
 
     public function configurePackage(Package $package): void
     {
-        $package
-            ->name('filament-spark-pie')
-            ->hasViews();
+        /*
+         * This class is a Package Service Provider
+         *
+         * More info: https://github.com/spatie/laravel-package-tools
+         */
+        $package->name(static::$name);
+
+        if (file_exists($package->basePath('/../resources/views'))) {
+            $package->hasViews(static::$viewNamespace);
+        }
+    }
+
+    public function packageRegistered(): void
+    {
+    }
+
+    public function packageBooted(): void
+    {
+        // Asset Registration
+        FilamentAsset::register(
+            $this->getAssets(),
+            $this->getAssetPackageName()
+        );
+    }
+
+    protected function getAssetPackageName(): ?string
+    {
+        return 'filament-spark-pie';
+    }
+
+    /**
+     * @return array<Asset>
+     */
+    protected function getAssets(): array
+    {
+        return [
+            Js::make('spark-pie-scripts', __DIR__ . '/../resources/dist/js/spark-pie.js'),
+        ];
     }
 }
